ml/requirements.txt
torch
torchvision
torchaudio
ultralytics
opencv-python
matplotlib
pyyaml
scipy
pandas
pillow
pillow-heif
labelImg